+++ app/Modules/Permissions/Database/Seeds/PermissionsTableSeeder.php
@@ -5,12 +5,12 @@ namespace App\Modules\Permissions\Database\Seeds;
 use Nova\Database\ORM\Model;
 use Nova\Database\Seeder;
 
-use App\Modules\Platform\Database\CreatePermissionsTrait;
+use App\Modules\Platform\Database\ManagePermissionsTrait;
 
 
 class PermissionsTableSeeder extends Seeder
 {
-    use CreatePermissionsTrait;
+    use ManagePermissionsTrait;
 
 
     /**
